CTPreparationgView Push

+++ CodingTestTimer/Domain/Model/CTSetting.swift
@@ -8,6 +8,6 @@
 import UIKit
 
 struct CTSetting {
-    var timeLimit: Int
+    var timeLimit: Double
     var problems: [Problem]
 }
